Move html.js
